Add support for multiple hostnames
The sink could previously only take a single hostname to connect to,
this change adds a new field on the property taking multiple hostnames.
If set, this list of hostnames will be used instead of the singular one,
passing it to the RabbitMQ ConnectionFactory's CreateConnection method
and allowing multiple hostnames to be used.

The IConnectionFactory interface for some reason does not contain the
version of CreateConnection methods that take a list of AmqTcpEndpoint
instances, which means the port cannot be customised on a per-host
basis. Changing RabbitMQClient.GetConnectionFactory to return a
ConnectionFactory rather than IConnectionFactory (and then also the
_connectionFactory field) should remedy that.